Move parsePackageSpec into core
