+++ packages/core/src/index.ts
@@ -1,3 +1,4 @@
 export * from "./checkPackage.js";
 export * from "./problems.js";
 export type * from "./types.js";
+export * from './parsePackageSpec.js';
