+++ packages/core/src/parsePackageSpec.ts
@@ -1,19 +1,13 @@
 import validatePackgeName from "validate-npm-package-name";
+import type { Failable, ParsedPackageSpec } from "./types.js";
 
 // Good grief https://semver.org/#is-there-a-suggested-regular-expression-regex-to-check-a-semver-string
 const semverRegex =
   /^(0|[1-9]\d*)\.(0|[1-9]\d*)\.(0|[1-9]\d*)(?:-((?:0|[1-9]\d*|\d*[a-zA-Z-][0-9a-zA-Z-]*)(?:\.(?:0|[1-9]\d*|\d*[a-zA-Z-][0-9a-zA-Z-]*))*))?(?:\+([0-9a-zA-Z-]+(?:\.[0-9a-zA-Z-]+)*))?$/;
 
-interface ParsedPackageSpec {
-  packageName: string;
-  version: string | undefined;
-}
-
-type Failable<T> = { status: "error"; error: string } | { status: "success"; data: T };
-
 export function parsePackageSpec(input: string): Failable<ParsedPackageSpec> {
-  let packageName: string;
-  let version: string;
+  let packageName;
+  let version;
   let i = 0;
   if (input.startsWith("@")) {
     i = input.indexOf("/");
